requirements.txt
openai>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0
chromadb>=0.4.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
pypdf>=3.0.0
docx2txt>=0.8
transformers>=4.44,<5
huggingface-hub>=0.24,<1
sentence-transformers>=3.0,<4
pytest>=7.0.0
